Version 9.2.1
Fixed problems with table-definitions for the new caching tables
Fixed missing cache-refresh after comments and comment*ratings on
records (using plugin-to-table-maps)
Added option attachments.useFlowPlayer - FlowPlayer, even the newest
version is full with deprecated jQuery and JavaScript. Default of the
new option is 1, if in future you see problems because of FlowPlayer,
set it to 0 and turn it off. Replacement for FlowPlayer will follow in
next version
Changed minimal value and default of advanced.sessionCompressionLevel to
1, avoiding problems with 0-zipped data using MYSQL 5.5